Minor cleanup - move TermsFacet to own file
